\name{ROC.data.ex}
\alias{ROC.data.ex}
\docType{data}
\title{
Example data
}
\description{
This hypothetical dataset contains records of the outcome, the predictor, gender, and age from 241 participants.
}
\usage{data("ROC.data.ex")}
\format{
  A data frame with 241 observations on the following 4 variables.
  \describe{
    \item{\code{outcome}}{a numeric vector}
    \item{\code{predictor}}{a numeric vector}
    \item{\code{gender}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
  }
}
\examples{
data(ROC.data.ex)
## maybe str(ROC.data.ex) ; plot(ROC.data.ex) ...
}
\keyword{datasets}
